Add `mavis.test.models`
This combines `mavis.test.onboarding` and `mavis.test.mavis_constants`
in to a single file that contains all models that represent concepts in
Mavis.
