+++ mavis/test/pages/consent.py
@@ -2,8 +2,7 @@ from datetime import date
 from playwright.sync_api import Page, expect
 from typing import Final, Optional
 
-from ..mavis_constants import ConsentMethod, Programme, ConsentRefusalReason
-from ..onboarding import School
+from ..models import ConsentMethod, ConsentRefusalReason, Programme, School
 from ..step import step
