+++ tests/test_online_consent_doubles.py
@@ -2,7 +2,7 @@ from datetime import date
 
 import pytest
 
-from mavis.test.mavis_constants import Programme, ConsentRefusalReason
+from mavis.test.models import Programme, ConsentRefusalReason
 
 pytestmark = pytest.mark.consent
